requirements.txt
wtforms
flask
gunicorn
SQLAlchemy
Flask-SQLAlchemy
Flask-migrate
requests
psycopg2
pytest
Flask-bcrypt
cryptography
Flask-Login
Flask-WTF
marshmallow==3.0.0rc4
marshmallow-oneofschema
python-dateutil
Faker
pytz